simplify processing of plain action directives
Instead of using docutils transformations hack as done with table
directives, plain directives now just creates action nodes directly
(without any transformation or pending nodes).

This siplifies the code for plain variants of directives and allows more
information to be discovered (eg. for the purpose of rstsource module),
but it also changes structure of resulting node tree a bit:

* plain directives would never produce any pending nodes
* action nodes are not wrapped into paragraph node anymore

Unfortunatelly, it also:

* break auto id feature for plain output