Fix builder with newer Sphinx (now 1.6 is required)
The way to set the translator class changed in 1.6. Accessing
self.translator_class raises an AttributeError, and looking at the
code the way to set the translator class should be set_translator,
or apparently just setting the default translator class.

Unfortunately setting the default seems to work only with Sphinx
newer than 1.6, so raise the requirements.

Fix #41